requirements.txt
flask>=3.0.0
tweepy>=4.14.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.5.0
Pillow>=10.0.0
